Dynamic dropdown placeholders
'Choose...' now appears as a placeholder for each dropdown. If a item is chosen and the move is then changed, the placeholder will return. Placeholder is the first element of each array; perhaps not the best implementation but it works. Just make sure not to accept an order until all placeholders are removed via choices.